refactor: Split parser expressions.rs into modular structure
Refactored monolithic 580-line expressions.rs into organized module structure to improve maintainability and code organization.

## Changes

### Module Structure
- Removed: `crates/parser/src/parser/expressions.rs` (580 lines)
- Created: `crates/parser/src/parser/expressions/` directory with:
  - `mod.rs` (46 lines) - Main entry point and coordinator
  - `operators.rs` (292 lines) - All operator parsing logic
  - `literals.rs` (127 lines) - Literal value parsing
  - `special_forms.rs` (75 lines) - CAST, EXISTS, NOT
  - `functions.rs` (58 lines) - Function call parsing
  - `identifiers.rs` (39 lines) - Column/table reference parsing
  - `subqueries.rs` (41 lines) - Subquery expression parsing

### Functions Organized by Category

**Operators (292 lines):**
- Precedence-based parsing: OR, AND, additive, multiplicative, comparison
- Special operators: IN, NOT IN, BETWEEN, NOT BETWEEN
- LIKE, NOT LIKE with pattern matching
- Quantified comparisons: ALL, ANY, SOME
- Expression list parsing

**Literals (127 lines):**
- Number literals (integers, decimals)
- String literals
- Boolean literals (TRUE, FALSE)
- NULL literal
- Typed literals: DATE, TIME, TIMESTAMP
- INTERVAL literals with field specifications

**Special Forms (75 lines):**
- CAST expressions
- EXISTS/NOT EXISTS
- Unary NOT operator

**Functions (58 lines):**
- Function call syntax parsing
- Argument list handling
- COUNT(*) wildcard support

**Identifiers (39 lines):**
- Simple column references
- Qualified names (table.column)
- Wildcard coordination

**Subqueries (41 lines):**
- Parenthesized expressions
- Scalar subqueries
- SELECT detection

## Benefits

✅ **Better Organization**: Expression types clearly separated by category
✅ **Improved Maintainability**: Each module focused on specific parsing logic
✅ **Enhanced Testability**: Modules can be tested independently
✅ **Clearer Precedence Handling**: Operator precedence isolated in operators.rs
✅ **Easier to Extend**: Clear location for new expression types

## Testing

- ✅ All 357 parser tests pass
- ✅ No behavior changes - pure refactoring
- ✅ Parse trees identical before/after
- ✅ File sizes reasonable (39-292 lines per module)

## Comparison

- Before: 580-line file with mixed expression logic
- After: 7 focused modules totaling 678 lines (98 additional from imports/docs)
- Result: Better organized, easier to navigate and extend

Closes #183
