+++ crates/parser/src/parser/expressions/mod.rs
@@ -1,0 +1,46 @@
+use super::*;
+
+// Submodules
+mod operators;
+mod literals;
+mod identifiers;
+mod functions;
+mod special_forms;
+mod subqueries;
+
+impl Parser {
+    /// Parse an expression (entry point)
+    pub(super) fn parse_expression(&mut self) -> Result<ast::Expression, ParseError> {
+        self.parse_or_expression()
+    }
+
+    /// Parse primary expression (literals, identifiers, parenthesized expressions)
+    pub(super) fn parse_primary_expression(&mut self) -> Result<ast::Expression, ParseError> {
+        // Try to parse as a literal
+        if let Some(expr) = self.parse_literal()? {
+            return Ok(expr);
+        }
+
+        // Try to parse as a special form (CAST, EXISTS, NOT EXISTS)
+        if let Some(expr) = self.parse_special_form()? {
+            return Ok(expr);
+        }
+
+        // Try to parse as a function call
+        if let Some(expr) = self.parse_function_call()? {
+            return Ok(expr);
+        }
+
+        // Try to parse as an identifier (column reference or qualified name)
+        if let Some(expr) = self.parse_identifier_expression()? {
+            return Ok(expr);
+        }
+
+        // Try to parse as a parenthesized expression or scalar subquery
+        if let Some(expr) = self.parse_parenthesized()? {
+            return Ok(expr);
+        }
+
+        Err(ParseError { message: format!("Expected expression, found {:?}", self.peek()) })
+    }
+}
