+++ crates/parser/src/parser/expressions/subqueries.rs
@@ -1,0 +1,41 @@
+use super::*;
+
+impl Parser {
+    /// Parse a subquery: (SELECT ...)
+    /// Used for scalar subqueries and IN operator
+    pub(super) fn parse_subquery(&mut self) -> Result<ast::SelectStmt, ParseError> {
+        // Expect opening paren
+        self.expect_token(Token::LParen)?;
+
+        // Parse SELECT statement
+        let select_stmt = self.parse_select_statement()?;
+
+        // Expect closing paren
+        self.expect_token(Token::RParen)?;
+
+        Ok(select_stmt)
+    }
+
+    /// Parse scalar subquery or parenthesized expression
+    pub(super) fn parse_parenthesized(&mut self) -> Result<Option<ast::Expression>, ParseError> {
+        match self.peek() {
+            Token::LParen => {
+                self.advance(); // consume '('
+
+                // Check if this is a scalar subquery by peeking at next token
+                if self.peek_keyword(Keyword::Select) {
+                    // It's a scalar subquery: (SELECT ...)
+                    let select_stmt = self.parse_select_statement()?;
+                    self.expect_token(Token::RParen)?;
+                    Ok(Some(ast::Expression::ScalarSubquery(Box::new(select_stmt))))
+                } else {
+                    // Regular parenthesized expression: (expr)
+                    let expr = self.parse_expression()?;
+                    self.expect_token(Token::RParen)?;
+                    Ok(Some(expr))
+                }
+            }
+            _ => Ok(None),
+        }
+    }
+}
